[ysh refactor] Reorganize library/ directory
func_cpython.py is now the stuff we want to rewrite.
