+++ prebuilt/dynamic-deps/filter-translate.txt
@@ -11,8 +11,8 @@ frontend/match.py
 pgen2/grammar.py
 pgen2/pnode.py
 pylib/path_stat.py
+library/func_cpython.py
 ysh/builtin_json.py
-ysh/objects.py
 ysh/cpython.py
 ysh/expr_eval.py
 osh/bool_stat.py
