+++ prebuilt/dynamic-deps/filter-typecheck.txt
@@ -1,7 +1,7 @@
 __init__.py
 typing.py
+library/func_cpython.py
 ysh/builtin_json.py
-library/func_misc.py
 pylib/collections_.py
 mycpp/mylib.py
 tools/deps.py
